backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "alembic[standard]>=1.18.4",
    "fastapi[standard]>=0.129.0",
    "pwdlib[argon2]>=0.3.0",
    "pyjwt>=2.11.0",
    "python-multipart>=0.0.22",
    "sqlmodel>=0.0.33",
]
